Added lowercase "dynamic" to work with 6880x
The 6880x for some reason leads with an * and uses a lowercase dynamic in the output. Both are now accounted for in the script. 

This also allows it to work with Aruba (HP) 5400 switches. The output for Aruba isn't quite right and the labels are off but at least you get output.